move queryist utils into sqlserver package
